requirements.txt
streamlit
transformers>=4.30.0
torch>=2.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
python-pptx>=1.0.0
sentencepiece>=0.1.99
protobuf>=3.20.0
fpdf>=1.7.2
numpy>=1.21.0
tqdm>=4.65.0
filetype>=1.2.0
